@@ -3,23 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@sanity/block-content-to-react": "^2.0.7",
-    "@sanity/client": "^2.8.0",
-    "@sanity/image-url": "^0.140.22",
+    "@sanity/block-content-to-react": "^3.0.0",
+    "@sanity/client": "^6.12.3",
+    "@sanity/image-url": "^1.0.2",
     "@sanity/uuid": "^3.0.1",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "date-fns": "^2.21.3",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "date-fns": "^2.30.0",
     "phosphor-react": "^1.4.1",
     "prismjs": "^1.29.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-social-icons": "^5.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0",
+    "react-scripts": "5.0.1",
+    "react-social-icons": "^5.15.0",
     "sitemap": "^8.0.0",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^3.5.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -46,5 +46,22 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "engines": {
+    "node": ">=14.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-async-generator-functions": "^7.20.7",
+    "@babel/plugin-proposal-dynamic-import": "^7.18.6",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-syntax-export-namespace-from": "^7.8.3",
+    "@babel/plugin-syntax-flow": "^7.14.5",
+    "@babel/plugin-transform-async-generator-functions": "^7.27.1",
+    "@babel/plugin-transform-dynamic-import": "^7.27.1",
+    "@babel/plugin-transform-react-jsx": "^7.14.9",
+    "@testing-library/dom": ">= 7.21.4",
+    "@types/react": "^ 18.0.0",
+    "typescript": ">= 2.7"
   }
 }
